integrate the new clause database with the rest of the sat solver
