+++ crates/solver/src/clause_db/impls.rs
