+++ crates/solver/src/clause_db/iter.rs
