+++ crates/solver/src/clause_db/tests.rs
