+++ crates/solver/src/clause_db/words.rs
